Consolidated common staking balance functions under parent Address class + test cases additions
